Update group_users_call_statistics.py
Odin had changed the GET Users Stats method so that if no calls are made to the given user, it returns an empty list. This caused an error in the group_users_call_statistics since in the method we try to access a key in the dictionary and if this is an empty list, the script will error. I updated the method so that if get.users_stats returns an empty list, we assign it to a dict and assign 0s to the variables